# internal/api/openapi/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Ledger API",
    "description": "Programmable double-entry accounting engine. An append-only event log is the source of truth, with derived projections for balances and transactions. All writes flow through the universal submit endpoint (`POST /v1/ledgers/{ledger_id}/submit`).",
    "version": "1.0",
    "contact": {
      "name": "Ledger API"
    }
  },
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/ledgers": {
      "get": {
        "operationId": "LedgerService_ListLedgers",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListLedgersResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "pageToken",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "LedgerService"
        ]
      },
      "post": {
        "summary": "Create a ledger",
        "operationId": "LedgerService_CreateLedger",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Ledger"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateLedgerRequest"
            }
          }
        ],
        "tags": [
          "Ledgers"
        ]
      }
    },
...
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1CreateLedgerRequest": {
      "type": "object",
      "example": {
        "id": "demo-ledger",
        "bucketId": "_default",
        "metadata": {
          "environment": "development",
          "owner": "finops-team"
        }
      },
      "properties": {
        "id": {
          "type": "string"
        },
        "bucketId": {
          "type": "string"
        },
        "metadata": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
...
    "v1Ledger": {
      "type": "object",
      "example": {
        "id": "demo-ledger",
        "bucketId": "_default",
        "state": "ACTIVE",
        "metadata": {
          "environment": "development"
        },
        "createdAt": "2026-01-15T10:00:00Z"
      },
      "properties": {
        "id": {
          "type": "string"
        },
        "bucketId": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "features": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "metadata": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "sealedAt": {
          "type": "string",
          "format": "date-time"
        },
        "issuerAccounts": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "v1ListLedgersResponse": {
      "type": "object",
      "properties": {
        "ledgers": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Ledger"
          }
        },
        "nextPageToken": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "description": "JWT bearer token. Send as `Authorization: Bearer \u003ctoken\u003e`. The token's `ledgers` claim governs per-ledger access.",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}